separate merge_cascade
