+++ be/src/exec/vectorized/sorting/sorting.h
@@ -54,18 +54,117 @@ void append_by_permutation(Column* dst, const Columns& columns, const Permutatio
 void append_by_permutation(Chunk* dst, const std::vector<ChunkPtr>& chunks, const Permutation& perm);
 void append_by_permutation(Chunk* dst, const std::vector<const Chunk*>& chunks, const Permutation& perm);
 
-struct SortedRun;
-struct SortDescs;
+struct SortDesc {
+    int sort_order;
+    int null_first;
+};
+
+struct SortDescs {
+    std::vector<int> sort_orders;
+    std::vector<int> null_firsts;
+
+    SortDescs(const std::vector<int>& orders, const std::vector<int>& nulls)
+            : sort_orders(orders), null_firsts(nulls) {}
+
+    size_t num_columns() const { return sort_orders.size(); }
+
+    SortDesc get_column_desc(int col) const {
+        SortDesc desc;
+        DCHECK_LT(col, sort_orders.size());
+        desc.sort_order = sort_orders[col];
+        desc.null_first = null_firsts[col];
+        return desc;
+    }
+};
+
+struct SortedRun {
+    ChunkPtr chunk;
+    std::pair<size_t, size_t> range;
+
+    SortedRun() = default;
+    ~SortedRun() = default;
+    explicit SortedRun(ChunkPtr ichunk) : chunk(ichunk), range(0, ichunk->num_rows()) {}
+    SortedRun(ChunkPtr ichunk, std::pair<size_t, size_t> irange) : chunk(ichunk), range(irange) {}
+    SortedRun(ChunkPtr ichunk, size_t start, size_t end) : chunk(ichunk), range(start, end) {}
+    SortedRun(const SortedRun& rhs) : chunk(rhs.chunk), range(rhs.range) {}
+    SortedRun& operator=(const SortedRun& rhs) {
+        if (&rhs == this) return *this;
+        chunk = rhs.chunk;
+        range = rhs.range;
+        return *this;
+    }
+
+    size_t num_columns() const { return chunk->num_columns(); }
+    size_t num_rows() const { return range.second - range.first; }
+    const Column* get_column(int index) const { return chunk->get_column_by_index(index).get(); }
+    bool empty() const { return range.second == range.first; }
+    void reset() {
+        chunk->reset();
+        range = {};
+    }
+    ChunkUniquePtr clone_chunk() {
+        if (range.first == 0) {
+            return chunk->clone_unique();
+        } else {
+            ChunkUniquePtr cloned = chunk->clone_empty(num_rows() - range.first);
+            cloned->append(*chunk, range.first, num_rows());
+            return cloned;
+        }
+    }
+};
+
+// Merge two sorted cusor
+class MergeTwoCursor {
+public:
+    MergeTwoCursor(const SortDescs& sort_desc, std::unique_ptr<SimpleChunkSortCursor>&& left_cursor,
+                   std::unique_ptr<SimpleChunkSortCursor>&& right_cursor);
+
+    // Use it as iterator
+    // Return nullptr if no output
+    StatusOr<ChunkUniquePtr> next();
+    Status consume_all(ChunkConsumer output);
+    ChunkProbeSupplier& as_supplier() { return _chunk_supplier; }
+    std::unique_ptr<SimpleChunkSortCursor> as_chunk_cursor();
+
+private:
+    StatusOr<ChunkUniquePtr> merge_sorted_cursor_two_way();
+    bool move_cursor();
+
+    SortDescs _sort_desc;
+    SortedRun _left_run;
+    SortedRun _right_run;
+    std::unique_ptr<SimpleChunkSortCursor> _left_cursor;
+    std::unique_ptr<SimpleChunkSortCursor> _right_cursor;
+    ChunkProbeSupplier _chunk_supplier;
+    std::unique_ptr<SimpleChunkSortCursor> _output_cursor = nullptr;
+    bool _wait_for_left = true;
+    bool _wait_for_right = true;
+};
+
+// Merge multiple cursors in cascade way
+class MergeCursorsCascade {
+public:
+    MergeCursorsCascade() = default;
+    ~MergeCursorsCascade() = default;
+
+    Status init(const SortDescs& sort_desc, std::vector<std::unique_ptr<SimpleChunkSortCursor>>&& cursors);
+    bool is_data_ready();
+    bool is_eos();
+    ChunkUniquePtr try_get_next();
+    Status consume_all(ChunkConsumer consumer);
+
+private:
+    std::vector<std::vector<std::unique_ptr<SimpleChunkSortCursor>>> _level_cursors;
+    std::vector<std::unique_ptr<MergeTwoCursor>> _mergers;
+};
 
 // Merge algorithms
-Status merge_sorted_chunks_two_way(const SortDescs& descs, const SortedRun& left_run, const SortedRun& right_run,
-                                   Permutation* output);
 Status merge_sorted_chunks_two_way(const SortDescs& descs, const ChunkPtr left, const ChunkPtr right,
                                    Permutation* output);
-Status merge_sorted_cursor_two_way(const SortDescs& sort_desc, std::unique_ptr<ChunkCursor> left_cursor,
-                                   std::unique_ptr<ChunkCursor> right_cursor, ChunkConsumer output);
-
-Status merge_sorted_cursor_cascade(const SortDescs& sort_desc, std::vector<std::unique_ptr<ChunkCursor>>& cursors,
+Status merge_sorted_cursor_two_way(const SortDescs& sort_desc, std::unique_ptr<SimpleChunkSortCursor> left_cursor,
+                                   std::unique_ptr<SimpleChunkSortCursor> right_cursor, ChunkConsumer output);
+Status merge_sorted_cursor_cascade(const SortDescs& sort_desc,
+                                   std::vector<std::unique_ptr<SimpleChunkSortCursor>>& cursors,
                                    ChunkConsumer consumer);
 
 // Merge in rowwise
